Fix Crash on Route Change
When the route changes, we're getting a new Kernel.
However the Notestates are not cleared and keeping a reference
to the old kernel, thus resulting in an invalid reference when
subsequent NoteOns are sent.

We set the initialized state to false, so they'll be recreated
in the next process() call.

Dont mind the formatting, this is extracted from the following commit.