Fix Crash on deallocation of render resources
Fix a crash that will occur when headphones are unplugged
on certain iOS Devices.
This was introduced earlier when trying to fix memory leaks,
unfortunately we are running into a double-free situation with
the C objects now.
We're now instead tracking state of the kernel initialization
in order to avoid those memory leaks.