Add Dark Mode Support
Add Support for Dark Mode by moving from the Boolean value we
previously used to an enum specifying the 3 available modes to the
user (light, dark, system).
Added a convenience function that checks the trait collection for
the current mode and falls back to false if not supported on OS.
However we should not run into users using "system" on versions below
iOS 13 because the option (segment) will be removed programatically
for these users.